Split session click-to-commit browser coverage
Moves the DEBT-378 session click-to-commit tests into a dedicated browser spec so the existing session question-flow spec stays below the nursery file-size warning threshold. Behavior is unchanged; this restores the lint warning count to the known baseline.
